+++ Genius/Views/FormView.swift
@@ -3,12 +3,13 @@
 // See LICENSE.txt for license information.
 //
 
+import _Concurrency
 import Foundation
 import SwiftUI
 
 struct FormView<ValueLabel: View, Value: FormValue>: View {
 
-	@State var content: CustomKeyValuePairs<LocalizedStringResource, CustomKeyValuePairs<ValueLabel, Value?>>?
+	@State var content: CustomKeyValuePairs<LocalizedStringResource, CustomKeyValuePairs<ValueLabel, Value??>>
 
 	let contentData: KeyValuePairs<LocalizedStringResource, KeyValuePairs<ValueLabel, any SystemInformationProtocol<Value>>>
 
@@ -17,44 +18,49 @@ struct FormView<ValueLabel: View, Value: FormValue>: View {
 		content: KeyValuePairs<LocalizedStringResource, KeyValuePairs<ValueLabel, any SystemInformationProtocol<Value>>>,
 	) {
 		self.contentData = content
+		self.content = content
+			.map { key, value in
+				(key: key, value: value.map { key, value in
+					(key: key, value: value.syncUIRepresentation)
+				})
+			}
 	}
 
 	var body: some View {
-		Group {
-			if let content {
-				Form {
-					ForEach(enumeratingID: content) { sectionTitle, sectionEntries in
-						if sectionEntries.contains(where: { $0.value != nil }) {
-							Section(sectionTitle) {
-								ForEach(enumeratingID: sectionEntries) { valueLabel, value in
+		Form {
+			ForEach(enumeratingID: content) { sectionTitle, sectionEntries in
+				if sectionEntries.contains(where: { $0.value != nil }) {
+					Section(sectionTitle) {
+						ForEach(enumeratingID: sectionEntries) { valueLabel, value in
+							if let value {
+								LabeledContent {
 									if let value {
-										LabeledContent {
-											value.formView
-										} label: {
-											valueLabel
-										}
+										value.formView
+									} else {
+										ProgressView()
+											.controlSize(.small)
 									}
+								} label: {
+									valueLabel
 								}
 							}
 						}
 					}
 				}
-				.formStyle(.grouped)
-				.frame(width: 512)
-			} else {
-				ProgressView()
 			}
 		}
+		.formStyle(.grouped)
+		.frame(width: 512)
 		.task {
-			let values: [[Value?]] = await contentData
-				.map { $0.value.map(\.value) }
-				.concurrentMap { await $0.concurrentMap { await $0.uiRepresentation } }
-			content = zip(contentData, values)
-				.map { keyValuePair, values in
-					(key: keyValuePair.key, value: zip(keyValuePair.value, values).map { keyValuePair, value in
-						(key: keyValuePair.key, value: value)
-					})
+			for sectionIndex in content.indices {
+				for entryIndex in content[sectionIndex].value.indices {
+					guard case .some(nil) = content[sectionIndex].value[entryIndex].value else { continue }
+					Task {
+						content[sectionIndex].value[entryIndex].value =
+							await contentData[sectionIndex].value[entryIndex].value.uiRepresentation.map(\.self)
+					}
 				}
+			}
 		}
 	}
 }